run TestWriteFailure against all versions (#1203)
* run TestWriteFailure against all versions

* Build independently tests that need environment variables

* Move ccm dependant tests under integration tag, add cassandra build tag, remove unit build tag
