+++ cassandra_test.go
@@ -1,4 +1,4 @@
-// +build all integration
+// +build all cassandra
 
 package gocql
 
@@ -21,59 +21,6 @@ import (
 
 	inf "gopkg.in/inf.v0"
 )
-
-// TestAuthentication verifies that gocql will work with a host configured to only accept authenticated connections
-func TestAuthentication(t *testing.T) {
-
-	if *flagProto < 2 {
-		t.Skip("Authentication is not supported with protocol < 2")
-	}
-
-	if !*flagRunAuthTest {
-		t.Skip("Authentication is not configured in the target cluster")
-	}
-
-	cluster := createCluster()
-
-	cluster.Authenticator = PasswordAuthenticator{
-		Username: "cassandra",
-		Password: "cassandra",
-	}
-
-	session, err := cluster.CreateSession()
-
-	if err != nil {
-		t.Fatalf("Authentication error: %s", err)
-	}
-
-	session.Close()
-}
-
-//TestRingDiscovery makes sure that you can autodiscover other cluster members when you seed a cluster config with just one node
-func TestRingDiscovery(t *testing.T) {
-	cluster := createCluster()
-	cluster.Hosts = clusterHosts[:1]
-
-	session := createSessionFromCluster(cluster, t)
-	defer session.Close()
-
-	if *clusterSize > 1 {
-		// wait for autodiscovery to update the pool with the list of known hosts
-		time.Sleep(*flagAutoWait)
-	}
-
-	session.pool.mu.RLock()
-	defer session.pool.mu.RUnlock()
-	size := len(session.pool.hostConnPools)
-
-	if *clusterSize != size {
-		for p, pool := range session.pool.hostConnPools {
-			t.Logf("p=%q host=%v ips=%s", p, pool.host, pool.host.ConnectAddress().String())
-
-		}
-		t.Errorf("Expected a cluster size of %d, but actual size was %d", *clusterSize, size)
-	}
-}
 
 func TestEmptyHosts(t *testing.T) {
 	cluster := createCluster()
@@ -2467,64 +2414,6 @@ func TestTokenAwareConnPool(t *testing.T) {
 	// TODO add verification that the query went to the correct host
 }
 
-func TestCustomPayloadMessages(t *testing.T) {
-	cluster := createCluster()
-	session := createSessionFromCluster(cluster, t)
-	defer session.Close()
-
-	if err := createTable(session, "CREATE TABLE gocql_test.testCustomPayloadMessages (id int, value int, PRIMARY KEY (id))"); err != nil {
-		t.Fatal(err)
-	}
-
-	// QueryMessage
-	var customPayload = map[string][]byte{"a": []byte{10, 20}, "b": []byte{20, 30}}
-	query := session.Query("SELECT id FROM testCustomPayloadMessages where id = ?", 42).Consistency(One).CustomPayload(customPayload)
-	iter := query.Iter()
-	rCustomPayload := iter.GetCustomPayload()
-	if !reflect.DeepEqual(customPayload, rCustomPayload) {
-		t.Fatal("The received custom payload should match the sent")
-	}
-	iter.Close()
-
-	// Insert query
-	query = session.Query("INSERT INTO testCustomPayloadMessages(id,value) VALUES(1, 1)").Consistency(One).CustomPayload(customPayload)
-	iter = query.Iter()
-	rCustomPayload = iter.GetCustomPayload()
-	if !reflect.DeepEqual(customPayload, rCustomPayload) {
-		t.Fatal("The received custom payload should match the sent")
-	}
-	iter.Close()
-
-	// Batch Message
-	b := session.NewBatch(LoggedBatch)
-	b.CustomPayload = customPayload
-	b.Query("INSERT INTO testCustomPayloadMessages(id,value) VALUES(1, 1)")
-	if err := session.ExecuteBatch(b); err != nil {
-		t.Fatalf("query failed. %v", err)
-	}
-}
-
-func TestCustomPayloadValues(t *testing.T) {
-	cluster := createCluster()
-	session := createSessionFromCluster(cluster, t)
-	defer session.Close()
-
-	if err := createTable(session, "CREATE TABLE gocql_test.testCustomPayloadValues (id int, value int, PRIMARY KEY (id))"); err != nil {
-		t.Fatal(err)
-	}
-
-	values := []map[string][]byte{map[string][]byte{"a": []byte{10, 20}, "b": []byte{20, 30}}, nil, map[string][]byte{"a": []byte{10, 20}, "b": nil}}
-
-	for _, customPayload := range values {
-		query := session.Query("SELECT id FROM testCustomPayloadValues where id = ?", 42).Consistency(One).CustomPayload(customPayload)
-		iter := query.Iter()
-		rCustomPayload := iter.GetCustomPayload()
-		if !reflect.DeepEqual(customPayload, rCustomPayload) {
-			t.Fatal("The received custom payload should match the sent")
-		}
-	}
-}
-
 func TestNegativeStream(t *testing.T) {
 	session := createSession(t)
 	defer session.Close()
@@ -2700,24 +2589,6 @@ func TestJSONSupport(t *testing.T) {
 	}
 	if state != "TX" {
 		t.Errorf("got state %q expected %q", state, "TX")
-	}
-}
-
-func TestUDF(t *testing.T) {
-	session := createSession(t)
-	defer session.Close()
-
-	if session.cfg.ProtoVersion < 4 {
-		t.Skip("skipping UDF support on proto < 4")
-	}
-
-	const query = `CREATE OR REPLACE FUNCTION uniq(state set<text>, val text)
-	  CALLED ON NULL INPUT RETURNS set<text> LANGUAGE java
-	  AS 'state.add(val); return state;'`
-
-	err := session.Query(query).Exec()
-	if err != nil {
-		t.Fatal(err)
 	}
 }
