+++ compressor_test.go
@@ -1,5 +1,3 @@
-// +build all unit
-
 package gocql
 
 import (
